# compiler: make `injectdestructors` a MIR pass