… compiler/mir/analysis.nim
… compiler/sem/injectdestructors.nim
… compiler/sem/mirexec.nim
… compiler/sem/optimizer.nim
… tests/arc/topt_refcursors.nim
… tests/arc/topt_wasmoved_destroy_pairs.nim
… tests/compiler/tmir_exec.nim
… tests/lang_objects/destructor/mhelper.nim
… tests/lang_objects/destructor/teval_order.nim
… tests/lang_objects/destructor/tglobals.nim
… tests/lang_objects/destructor/tlate_use.nim
… tests/lang_objects/destructor/tlvalue_conversions.nim
… tests/lang_objects/destructor/tobjfield_analysis.nim
… tests/lang_objects/destructor/tself_assign.nim
… tests/lang_objects/destructor/ttuple_unpacking.nim
… tests/lang_objects/destructor/tv2_cast.nim
… tests/lang_objects/destructor/tview_ownership.nim