@saem 2022-10-17
I think I understand it, going to write that out for confirmation:

injectdestructors:
1. has mirgen generates an initial `MirTree` (`genProcBody`)
2. then comes up with a list of changes it's going to make (`analyse`)
3. applies them to the `MirTree`
4. converts the updated `MirTree` to `PNode`

----

Prior to reading it, what I was looking for:
- that it's basically all lisp-y on the inside, the AST is a program (we briefly discussed it before in chat and it turns out we're very much on the same page)
- translation to any intermediate should "remember" where it came from
- the preivous point is most true wrt to changes made for precise error/warnings (need lineage)

It seems we're doing the above for the most part is all present, where we have the `ChangeSet` with `Action` remember the `PNode` it's acting upon. So errors should be able to trace their lineage back. The second bullet point I'm a little less sure about because I can't see the how we could correlate original `PNode` to initial `MirNode`.

So I'm not sure how we flow the info back up to Sem n order for us to ensure we can flow constraints (mostly types info) down from Sem and warnings/errors telemetry backup.