@zerbina 2022-10-17
> I think I understand it, going to write that out for confirmation:

Your understanding is correct. I apologize for not having provided a high-level overview.

---

I think I insufficiently described my motivation for this PR, and since that information can help in understanding some decisions behind the current implementation, here goes a second attempt:

The main contributors to my decision of working on this PR first are as following:
* the pre-PR `injectdestructors` produces AST that violates various (reasonable) expectations of `irgen`. For example, sometimes a produced `nkStmtListExpr` has no type (i.e. it isn't actually an expression) or it has a type but doesn't end in an expression. While possible to work around these things, additional logic would be required and `irgen` would also have to become less strict about it's input.
* the idea of implementing the move-analyser at the back-end IR level has been around since the new back-end's inception, but I no longer think that it's a good idea. Over time, the additional complexity of having the back-end IR accommodate for the move-analyser started to outweigh the benefits -- the ideas I collected on how the move-analyser could be implemented work better (if at all) with a slightly higher-level IR.
* `irgen` is quite complex. It does a lot of different things, like performing general transformations and lowering some magics. Then there's also the deferred type translation plus the subsequent fix-up of all type IDs required by it.

So to rephrase, the goal here is to introduce an IR that:
* is meant to support `injectdestructors`, `lambdalifting`, and, later on, maybe also other parts of `transf` (e.g. `closureiterators`, although this one will probably become unnecessary with the introduction of CPS transforms) efficiently operating on it
* works as the input to `irgen`

My current plan for the how the flow should look like in the near-term:
```
input ->
  transform() ->               # `transf`
    toMir() ->                 # currently `genProcBody`
      lambdalifting() ->
        injectdestructors() ->
          backend()            # `irgen`
``` 

Using `MirNode` in it's current form as the input to new back-end (or more specifically, `irgen`) would have the following benefits already:
* multiple transformations previously performed during `irgen` (listed in the PR's description) are now performed in `mirgen`. This would simplify `irgen` a lot.
* the elision of nodes irrelevant to the back-end (e.g. `nkBindStmt`, `nkTypeDef`, etc) together with the linear tree structure make scanning of dependencies (types, procedures, constants, literals, etc.) trivial, rendering, among other things, deferred type translation for `irgen` obsolete.

The main reason behind the `MirNode` -> `PNode` translation is to keep the current code-generators working without having to keep two implementations of `injectdestructors` around.

---

> The second bullet point I'm a little less sure about because I can't see the how we could correlate original `PNode` to initial `MirNode` -- the final hop.

A `MirNode`'s origin is currently not kept track of (maybe to a source location, but not to a `PNode`) - this is something that's still missing.

> So I'm not sure how we flow the info back up to Sem n order for us to ensure we can flow constraints (mostly types info) down from Sem and warnings/errors telemetry backup.

I'm not sure I understand. In which case is it required for information to flow back from the MIR to Sem? Sorry if it's obvious.